compute color max values correclty
The max color values in screen->serverFormat were populated with wrong
values on some platforms (i.e. some NVIDIA cards), causing invalid
color translation by libvncserver.

The cause of the problem is invalid `bits_per_rgb_value` in visual
info, which was set to 11 in my own case. x11vnc ignores this value, so
it should be fair to ignore this in krfb, too.